Fix rebase fallout
