+++ packages/sdk/src/gql.ts
@@ -1,0 +1,3 @@
+import noop from "tagged-template-noop";
+
+export const gql = noop;
