+++ packages/sdk/tsconfig.json
@@ -4,7 +4,8 @@
     "composite": true,
     "outDir": "./dist",
     "resolveJsonModule": true,
-    "rootDir": "./src"
+    "rootDir": "./src",
+    "baseUrl": "./src"
   },
   "include": ["./src"]
 }
